requirements.txt
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-rest-swagger==2.2.0
Pillow==8.3.1
python-decouple==3.4
